Add Nix build environment.
Part of preservation of build environment. Another solution exists.

There is issue with build with new binutils (started from 2.35 tested with 2.37).
For example:
ld: cannot use executable file 'bin/3c90x.pxe.rt1.uo' as input to a link
which is because there is just change in binutils.

There is better fix in makefile but this build env can be handy.